Allow to control version fields
